requirements.txt
mockito
pyserial
protobuf
amber-python-clients